Refactored shell to make use of the new AbstractCompositeShell
