+++ NaluMaterialSimpleApplication/src/main/java/de/gishmo/gwt/example/nalu/simpleapplication/client/ui/shell/Shell.ui.xml
